[FIRRTL] Split intrinsic op into expr vs stmt.
Expression form is Pure and must have a result.
Optionally carries a name (like other expressions).

FIRRTL operations are expected to be one of expression or statement
(see Visitor, firrtl::isExpression, so on) so split the combined
operation into these two kinds.

Statement form is not pure and does not have a result.