[Moore] Remove formatting of types as SV string
Remove the `format` functions from the types in the Moore dialect. These
could format a type as SystemVerilog syntax to a certain extent, but
they are unused in practice. The formatting was useful in the original
Rust implementation to produce nice error messages during type checking.
Since Slang performs type checking for us, the formatting is no longer
needed.