[FIRRTL] AnnoTarget: use LLVM style casts
This changes AnnoTargets to use LLVM style casts instead of "trailing"
member functions. The LLVM style casts are now usable with AnnoTargets
with improvements to the upstream framework.  The current style of cast
used with AnnoTargets will no longer be supported by TypeSwitch, so this
PR switches to the recommended idiom.