[ExportVerilog] Emit extra parentheses around reduction ops (#3357)
This closes #3001. This PR is a follow-up onto
reduction operators appear in the head of subexpression such as `a & b & &a & a`.
This PR fixes the issue by setting precedence of reduction ops to lowest.

Example:
```mlir
hw.module @foo(%a: i4, %b: i1) -> (o1:i1) {
  %one4 = hw.constant -1 : i4

  %and1 = comb.icmp eq %a, %one4 : i4
  %and2 = comb.icmp eq %a, %one4 : i4
  %and3 = comb.icmp eq %a, %one4 : i4
  %and = comb.and %and1, %b, %and2, %and3  : i1
  hw.output %and : i1
}
```

`circt-opt -export-verilog` produces:

```verilog
module Foo(
  input  [3:0] a,
  input        b,
  output       o1);

  assign o1 = (&a) & b & (&a) & (&a);
endmodule
```