[FIRRTL] Ensure hierpath considers owning module in LowerClasses.
We already had logic to detect if the owning module was in the middle
of a hierarchical path, and build up the prefix just to the owning
module in this case. However, we still used the entire original
hierarchical path in the new path, which caused the prefix to the
owning module to be duplicated. This is fixed by not only setting the
moduleName to the owning module, but also trimming the prefix to the
owning module in the original path.