[Moore] SymbolVisibility attribute printing support for SVModuleOp
We can append the appropriate symbol visibility to svmoduleOp according
to the structures of instanceSymbol provided by the slang frontend. slang
provides function to get the root of the design. Calling this method
could get all top-level instanceSymbols and help determine which ones
should be tagged. Note that the visibility attribute now used does not
contain the `nested`.