[ESI] Add verify connections pass
New pass to verify that channels and bundles are used the appropriate
number of times. The wrap/unwrap, pack/unpack operations already have
these checks, but when a channel/bundle is produced by a module (or
some other op), it is not checked. A pass is the best way I could figure
out how to implement the check.

Fixes #7286.