[FIRRTL] AssignOutputDirs: Ensure absolute dirs end with trailing slash
When making a directory absolute, the remove_dots helper canonicalizes away any
trailing slash. Add the slash back on, which ensures that we interpret the path
as a directory correctly.

This fixes a bug where a path, when LCA'd with itself, drops the last segment.

An example of the problem: given a path foo/bar/, if we drop the trailing
slash, foo/bar is interpreted as a file called bar in a directory called foo.
When we LCA foo/bar with itself, the common parent directory is determined to
be foo/. By retaining the trailing slash, we can ensure that the correct LCA
path foo/bar/ is computed.

Fixes: #7347