[OM] Pass Python values back and forth, not Attributes. (#7417)
Internally, primitive OM EvaluatorValues are represented as
TypedAttributes. This is great internally, but when we pass these from
C++ out to Python, we have to use a very inefficient method to pull
the Python value out of the attribute.

This updates how primitives are handled at the Python <> C++ interface
to directly construct the appropriate Python values and return
them. Similarly, for top-level inputs to the Evaluator, Python values
are directly accepted and converted to Attributes internally.

On large designs, this was shown to decrease single threaded CPU time
to process large amounts of OM data by roughly 70%. There is no
difference in the output.