[ESI][Runtime] Poll method and optional service thread polling
Add a poll method to ports, a master poll method to the Accelerator, and
the ability to poll from the service thread. Also, only spin up the
service thread if it's requested.

The service thread polling (in particular) required some ownership
changes: Accelerator objects now belong to the AcceleratorConnection so
that the ports aren't destructed before the service thread gets shutdown
(which causes an invalid memory access). This particular binding isn't
ideal, is brittle, and will be an issue for anything doing the polling.
Resolving #7457 should mitigate this issue.

Backends are now _required_ to call `disconnect` in their destructor.