[FIRRTL] Don't capture non-passives in LowerLayers
Fix a bug where the LowerLayers pass could create output ports on a module
lowered from a bind convention layer.  Avoid this entirely by moving all
subfield, subindex, and subaccess operations (sub-* ops) out of
layerblocks before modules are created.

LowerLayers works by converting "captured" values into ports.  However,
the sub-* ops may capture non-passive types allowably, but then never
drive them.  E.g., consider the following:

    %0 = firrtl.wire : !firrtl.bundle<a : uint<1>, b flip: uint<1>>
    firrtl.layerblock @A {
      %1 = firrtl.subfield %0[a] : !firrtl.bundle<a : uint<1>, b flip: uint<1>>
      %2 = firrtl.node %1 : !firrtl.uint<1>
    }

Naively, this "captures" the non-passive %0.  However, this is really only
capturing the _passive portion_ of %0 through a subfield.  Without this
commit, LowerLayers will try to create a port with the same type as %0
when it should be creating a port with the same type of %1.  In order to
determine what the port is, LowerLayers needs to know what is actually
captured and not blindly assume that anything captured needs to be a port.

Because this analysis may be tricky, instead solve this by moving the
sub-* ops outside the layerblock before computing captures.  The captures
can then _continue_ to be naively computed by seeing if a value is defined
outside the layerblock.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>