[FIRRTL] Use new layer ABI
Align CIRCT with the finalized layer ABI [[1]].  This changes the file
name of the bind file used to enable a layer.  This specifically switches
from using underscore as a delimiter to a hyphen.  This avoids a problem
where a circuit that contained modules or layers whose names contained
underscores could result in multiple layer enable files that had, by the
ABI, the exact same name.

[1]: https://github.com/chipsalliance/firrtl-spec/pull/233

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>