[ESI][Runtime] Logging API
The ESI logging API is intended to be agnostic of the application's
logging framework. In other words, it can be adapted to the
application's logging method with ease -- users must only extend one
class. It also comes with a simple text logger.

It is also designed to enable debug logging in Release builds with
(very) minimal performance overhead if the user does not set the log
level to debug.

This PR introduces the API, but doesn't use it much. More plumbing will
be required to actually use it. In the interest of keeping this PR
small, this is left for future work.